Introduce new shipperMeta files and logic
We want a new logic around shipped blocks that tracks when were they
shipped. To do this we have modified the logic to both read and write
the shipper meta files where we will try first to read
mimir.shipper.json and if its not possible then thanos.shipper.json.

Also the logic keeps thanos.shipper.json up to date in case of a
possible rollback.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>